feat: Enable third party app use of http server
Migrating the entry point `package main` to `cmd` enables using the http
server package in other applications.
