+++ README.md
@@ -10,3 +10,4 @@ Example implementations of common requirements for website serving applications.
 1. GitHub Action CI to lint & test - https://github.com/karlskewes/go-yahs/pull/2
 1. Graceful shutdown - https://github.com/karlskewes/go-yahs/pull/4
 1. Testable application invocations. Split `main()` with `Run()` - https://github.com/karlskewes/go-yahs/pull/5
+1. Enable importing into other applications, move `package main` to `cmd/..` - https://github.com/karlskewes/go-yahs/pull/6
